Increase timeout in timeout test
The test_playbook test in the Ansible tests (which tests a bunch of
interactions between Zuul and Ansible) has a test job called "timeout"
which verifies that everything works if a job timeout hits while
ansible is running.  We currently allow the job to run for 5 seconds,
but that includes both pre-run and run playbooks.  Sometimes the job
runs slowly and we hit the timeout during the pre-run playbook, which
means it gets retried, which the test does not expect.  Double our
allowance and let the job run for 10 seconds before hitting the timeout
so that it's more likely it hits during the run playbook.

Change-Id: I396647117f7362c4b547ed374292c12bd5a73435