Cache getUser in Github connection
We currently request the user of the sender login for each incoming
event. This only contains static information so this can be easily
cached to reduce the amount of api requests towards github.

Log analysis shows that in our deployment around 10% of the api
requests are done to request users where a big portion of that are
duplicated (e.g. zuul itself). So this should show a very high cache
hit rate.

Change-Id: Ia6ab97ecdfad78e76b31ede5e8080407df6bc84d