Provide some documentation for the checks API implementation
There were some questions raised in IRC why we chose to implement the
checks API this way and why we don't support e.g. check runs on job
level.

This adds a chapter to the discussion section of the documentation
providing more detailed information about the concepts behind the
implementation and why certain decisions where made.

Change-Id: I750589a3589a624ac0b38006d030817e851f2318