Render buildset progress bar correctly
The buildset progress bar was not rendered using the same colors as the
job statuses itself. This happened because the job results were
interpreted differently for the job status and the progress bar.

There is now a common jobStrResult() function to have the same result
string in both cases. Missing result cases were added as needed.

Change-Id: I202959766e193efacd00b0dc85bf5160f8412a24