Tune automatic garbage collection of git repos
Git has an automatic garbage collection mechanism that is intended to
keep operations efficient when working with a git repository. However
the default settings are tuned to work best with repos that are
direcly used by humans. In zuul we see a slowly degrading performance
of job startup times over time which we fixed by pruning caches every
two weeks.

We found out that instead of pruning the caches we can have a similar
effect by tuning the automatic garbage collection settings. The effect
was that after a restart with optimized settings our worst case job
startup times dropped from around 15 minutes to about 9 minutes.

Change-Id: I594ceff9073f7ca14221d4f37716c76b462fc3d9