Don't retry after Gerrit HTTP 409s
Gerrit will return an HTTP status 409 when you try to modify changes
that are already merged or abandoned. For example if you try to vote on
a change in these closed states. In those cases we don't actually want
to retry the request as doing so is a waste of effort.

Update the gerritconnection drive to handle HTTP 409s as a special case
that should not be retried.

Change-Id: I4bd7b8e5aa512bf0ba1d1bf7dd9a39c3843e6ff6