Fix cherrypy dependency conflict
Current set of dependencies brings latest cherrypy which is
incompatible with cheroot we need. This pins cherrypy to last
version that is compatible. The current verision of pip will
still install cherrypy, but the upcoming depsolver is unable to
resolve the conflict.

We also add a fail-safe `pip check` that runs the new depsolver
and should prevent a similar regressions from appearing.

Change-Id: I2b0fa43b906d0919854d30da0ee825002444b201