Fix test_crd_check_unknown tests
These tests have a race where we rely on input events for unknown
projects in gerrit/github to populate our fake change and project
databases for those event sources. Before our dependent change is able
to be processed we need the parent change's event to be fully processed
and populate the database.

Without this we get database lookup errors which leads to us never
running jobs for the dependent change. When that happens we are unable
to properly assert the test conditions that the dependent change is
tested without the parent change.

Fix this by adding the parent change events, waiting for things to
settle, then adding the child change events.

Change-Id: I0641c44051d760c7be0f408f6b9b59222b8d4e90