Simplify FakeGithubClient and FakeGithubSession
These two fakes are related to each other. Before this change
FakeGithubClient would create a FakeGithubSession and FakeGithubSession
would create FakeGithubClients. THis recursion is confusing and
unnecessary.

Under normal operation the FakeGithubClient exists first so we pass it
into its FakeGithubSessions and the fake session is able to use this
client object rather than recursively creating a new client.

Change-Id: I9ff52061d62e844ca3f2185ea752ca39962b2a9f